requirements.txt
mfrc522==0.0.7
mysql_connector_repackaged==0.3.1
picamera==1.13
python-dotenv==0.20.0
pyttsx3==2.90
